create user abm commands
