+++ app/commands/ping.js
