+++ app/index.js
@@ -1,6 +1,6 @@
 const { app, BrowserWindow, ipcMain } = require('electron');
 const { connect: databaseConnect } = require('./resources/database');
-require('./resources/ping');
+require('./commands/ping');
 
 // Keep a global reference of the window object, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
